+++ src/units/udisplay.cpp
@@ -599,7 +599,8 @@ void unit_die(const map_location& loc, unit& loser,
 void unit_attack(display * disp, game_board & board,
                  const map_location& a, const map_location& b, int damage,
                  const attack_type& attack, const_attack_ptr secondary_attack,
-                 int swing,const std::string& hit_text,int drain_amount,const std::string& att_text, const std::vector<std::string>* extra_hit_sounds)
+                 int swing,const std::string& hit_text,int drain_amount,const std::string& att_text, const std::vector<std::string>* extra_hit_sounds,
+                 bool attacking)
 {
 	if(do_not_show_anims(disp) || (disp->fogged(a) && disp->fogged(b)) || !preferences::show_combat()) {
 		return;
@@ -620,6 +621,13 @@ void unit_attack(display * disp, game_board & board,
 	assert(def.valid());
 	unit &defender = *def;
 	int def_hitpoints = defender.hitpoints();
+	const_attack_ptr weapon = attack.shared_from_this();
+	auto ctx = weapon->specials_context(attacker.shared_from_this(), defender.shared_from_this(), a, b, attacking, secondary_attack);
+	std::optional<decltype(ctx)> opp_ctx;
+
+	if(secondary_attack) {
+		opp_ctx.emplace(secondary_attack->specials_context(defender.shared_from_this(), attacker.shared_from_this(), b, a, !attacking, weapon));
+	}
 
 	att->set_facing(a.get_relative_dir(b));
 	def->set_facing(b.get_relative_dir(a));
@@ -640,16 +648,20 @@ void unit_attack(display * disp, game_board & board,
 	unit_animator animator;
 
 	animator.add_animation(attacker.shared_from_this(), "attack", att->get_location(), def->get_location(), damage, true, text_2,
-		(drain_amount >= 0) ? color_t(0, 255, 0) : color_t(255, 0, 0), hit_type, attack.shared_from_this(),
+		(drain_amount >= 0) ? color_t(0, 255, 0) : color_t(255, 0, 0), hit_type, weapon,
 		secondary_attack, swing);
 
 	// note that we take an anim from the real unit, we'll use it later
 	const unit_animation* defender_anim = def->anim_comp().choose_animation(*disp, def->get_location(), "defend",
-		att->get_location(), damage, hit_type, attack.shared_from_this(), secondary_attack, swing);
+		att->get_location(), damage, hit_type, weapon, secondary_attack, swing);
 
 	animator.add_animation(defender.shared_from_this(), defender_anim, def->get_location(), true, text, {255, 0, 0});
 
-	for(const unit_ability& ability : attacker.get_abilities_weapons("leadership", attack.shared_from_this(), secondary_attack)) {
+	unit_ability_list abilities = attacker.get_abilities_weapons("leadership", weapon, secondary_attack);
+	for(auto& special : attacker.checking_tags()) {
+		abilities.append(weapon->list_ability(special));
+	}
+	for(const unit_ability& ability : abilities) {
 		if(ability.teacher_loc == a) {
 			continue;
 		}
@@ -663,10 +675,10 @@ void unit_attack(display * disp, game_board & board,
 		leader->set_facing(ability.teacher_loc.get_relative_dir(a));
 		animator.add_animation(leader.get_shared_ptr(), "leading", ability.teacher_loc,
 			att->get_location(), damage, true,  "", {0,0,0},
-			hit_type, attack.shared_from_this(), secondary_attack, swing);
+			hit_type, weapon, secondary_attack, swing);
 	}
 
-	for(const unit_ability& ability : defender.get_abilities_weapons("resistance", secondary_attack, attack.shared_from_this())) {
+	for(const unit_ability& ability : defender.get_abilities_weapons("resistance", secondary_attack, weapon)) {
 		if(ability.teacher_loc == a) {
 			continue;
 		}
@@ -680,7 +692,7 @@ void unit_attack(display * disp, game_board & board,
 		helper->set_facing(ability.teacher_loc.get_relative_dir(b));
 		animator.add_animation(helper.get_shared_ptr(), "resistance", ability.teacher_loc,
 			def->get_location(), damage, true,  "", {0,0,0},
-			hit_type, attack.shared_from_this(), secondary_attack, swing);
+			hit_type, weapon, secondary_attack, swing);
 	}
 
 
@@ -717,7 +729,11 @@ void reset_helpers(const unit *attacker,const unit *defender)
 	display* disp = display::get_singleton();
 	const unit_map& units = disp->get_units();
 	if(attacker) {
-		for(const unit_ability& ability : attacker->get_abilities("leadership")) {
+		unit_ability_list attacker_abilities = attacker->get_abilities("leadership");
+		for(auto& special : attacker->checking_tags()) {
+			attacker_abilities.append(attacker->get_abilities(special));
+		}
+		for(const unit_ability& ability : attacker_abilities) {
 			unit_map::const_iterator leader = units.find(ability.teacher_loc);
 			assert(leader != units.end());
 			leader->anim_comp().set_standing();
@@ -725,7 +741,11 @@ void reset_helpers(const unit *attacker,const unit *defender)
 	}
 
 	if(defender) {
-		for(const unit_ability& ability : defender->get_abilities("resistance")) {
+		unit_ability_list defender_abilities = defender->get_abilities("resistance");
+		for(auto& special : defender->checking_tags()) {
+			defender_abilities.append(defender->get_abilities(special));
+		}
+		for(const unit_ability& ability : defender_abilities) {
 			unit_map::const_iterator helper = units.find(ability.teacher_loc);
 			assert(helper != units.end());
 			helper->anim_comp().set_standing();